Add SEO, LLM discovery, and component refactoring
- Add robots.ts, sitemap.ts for search engine crawling
- Add llms.txt and .well-known/ai-plugin.json for LLM discovery
- Add OpenGraph/Twitter metadata with canonical URLs
- Extract client-side demos into separate components for proper
  Next.js server/client component separation
- Add AI content declaration metadata

Refs AINative-Studio/core#1373
